finance-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
fastapi[security]
python-jose
bcrypt
psycopg2
passlib[bcrypt]
python-jose[cryptography]
python-multipart
